textxyz/title_classifier/requirements.txt
nltk
numpy
gensim==3.0.0
tensorflow_gpu==1.12.0